requirements.txt
asgiref==3.7.2
cffi==1.15.1
click==8.1.7
cryptography==41.0.3
Django==4.2.4
django-resized==1.0.2
django-rest-knox==4.2.0
djangorestframework==3.14.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
Pillow==10.0.0
platformdirs==3.10.0
psycopg2-binary==2.9.7
pycparser==2.21
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
